settup of the documentation generation using rosdoc_lite

+++ niryo_robot_hardware_stack/conveyor_interface/src/conveyor_interface_core.cpp
@@ -135,7 +135,7 @@ namespace ConveyorInterface {
         else if (req.cmd == conveyor_interface::SetConveyor::Request::REMOVE)
         {
             bool conveyor_found = false;
-            for(int i = 0; i < _list_conveyor_id.size(); i++)
+            for(size_t i = 0; i < _list_conveyor_id.size(); i++)
             {
                 if(req.id == _list_conveyor_id.at(i))
                 {
